Feat: convert expense entity String To double

+++ app/src/main/java/com/example/minhasdespesas/presentation/screens/expenseDetail/ExpenseBottomSheet.kt
@@ -168,7 +168,7 @@ fun ExpenseBottomSheet(
                 onClick = {
                     bottomSheetViewModel.saveNewExpense(
                         expenseName,
-                        expenseValue,
+                        expenseValue.toDouble(),
                         newCategoryName,
                         selectedColor,
                         expenseDetailViewModel.convertDateToTimestamp(date)
